backend/pyproject.toml
[project]
name = "ultraskate-buddy-api"
version = "0.1.0"
description = "Ultraskate Stats Viewer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "fastapi>=0.125.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
